+++ app.py
@@ -11,4 +11,4 @@ app = create_app()
 
 if __name__ == '__main__':
     from hemlock.app import socketio
-    socketio.run(app, debug=True)
+    socketio.run(app, host='0.0.0.0', debug=True)