+++ hemlock/qpolymorphs/__init__.py
@@ -6,6 +6,7 @@ from .choice import Choice, Option
 from .dashboard import Dashboard
 from .download import Download
 from .file import File
+from .hidden import Hidden
 from .input import Input
 from .label import Label
 from .range import Range, RangeInput, Slider, likert